Benchmarks for Mailbox
