+++ examples/MailboxBenchmark/MailboxBenchmark.csproj
@@ -7,4 +7,7 @@
     <ProjectReference Include="..\..\src\Proto.Actor\Proto.Actor.csproj" />
     <ProjectReference Include="..\..\src\Proto.Mailbox\Proto.Mailbox.csproj" />
   </ItemGroup>
+  <ItemGroup>
+    <PackageReference Include="BenchmarkDotnet" Version="0.10.8" />
+  </ItemGroup>
 </Project>
