+++ src/Proto.Mailbox/Mailbox.cs
@@ -24,7 +24,7 @@ namespace Proto.Mailbox
         void Start();
     }
 
-    internal static class BoundedMailbox
+    public static class BoundedMailbox
     {
         public static IMailbox Create(int size, params IMailboxStatistics[] stats)
         {
